refactor: move audio, downloader, progress packages to pkg
