+++ internal/server/types.go
@@ -4,7 +4,7 @@ import (
 	"time"
 
 	"github.com/jaki95/dj-set-downloader/internal/domain"
-	"github.com/jaki95/dj-set-downloader/internal/progress"
+	"github.com/jaki95/dj-set-downloader/pkg/progress"
 )
 
 // JobStatus represents the current state of a processing job
