+++ pkg/audio/interface.go
