+++ pkg/downloader/interface.go
